move simd tests out of ocaml
