+++ ocaml/testsuite/tests/unboxed-primitive-args/test_common.c
@@ -15,8 +15,6 @@
 
 #include <caml/mlvalues.h>
 #include <caml/bigarray.h>
-#include <smmintrin.h>
-#include <emmintrin.h>
 
 char *ocaml_buffer;
 char *c_buffer;
@@ -36,19 +34,4 @@ value test_cleanup_normal(void)
 double test_cleanup_float(void)
 {
   return 0.;
-}
-
-int64_t vec128_low_int64(__m128i v) 
-{
-  return _mm_extract_epi64(v, 0);
-}
-
-int64_t vec128_high_int64(__m128i v) 
-{
-  return _mm_extract_epi64(v, 1);
-}
-
-__m128i vec128_of_int64s(int64_t low, int64_t high) 
-{
-  return _mm_set_epi64x(high, low); 
 }
