+++ testsuite/tests/unboxed-primitive-args/common.ml
@@ -8,6 +8,7 @@ type 'a typ =
   | Int64     : int64     typ
   | Nativeint : nativeint typ
   | Float     : float     typ
+  | Vec128    : vec128    typ
 
 type 'a proto =
   | Ret : 'a typ -> 'a proto
@@ -44,6 +45,10 @@ let expand_test = function
     Test (s, fn, a ** b ** c ** d ** e ** f ** Ret g)
   | T (s, fn, p) -> Test (s, fn, p)
 
+external vec128_of_int64s : int64 -> int64 -> vec128 = "" "vec128_of_int64s" [@@noalloc] [@@unboxed] 
+external vec128_low_int64 : vec128 -> int64 = "" "vec128_low_int64" [@@noalloc] [@@unboxed]
+external vec128_high_int64 : vec128 -> int64 = "" "vec128_high_int64" [@@noalloc] [@@unboxed]
+
 let string_of : type a. a typ -> a -> string = function
   | Int       -> Int.to_string
   | Int32     -> Printf.sprintf "%ldl"
@@ -51,6 +56,8 @@ let string_of : type a. a typ -> a -> string = function
   | Nativeint -> Printf.sprintf "%ndn"
   | Float     ->
       fun f -> Printf.sprintf "float_of_bits 0x%LxL" (Int64.bits_of_float f)
+  | Vec128    -> 
+      fun v -> Printf.sprintf "vec128 %Ld:%Ld" (vec128_high_int64 v) (vec128_low_int64 v)
 
 let rec arity : type a. a proto -> int = function
   | Ret _ -> 0
@@ -59,7 +66,7 @@ let rec arity : type a. a proto -> int = function
 module Buffer = struct
   type t = (char, int8_unsigned_elt, c_layout) Array1.t
 
-  let arg_size = 8
+  let arg_size = 16
 
   let create ~arity : t =
     Array1.create char c_layout ((arity + 1) * arg_size)
@@ -75,6 +82,14 @@ module Buffer = struct
   external get_int64 : t -> int -> int64 = "%caml_bigstring_get64"
   external set_int32 : t -> int -> int32 -> unit = "%caml_bigstring_set32"
   external set_int64 : t -> int -> int64 -> unit = "%caml_bigstring_set64"
+
+  let get_vec128 buf ~arg = 
+    let low, high = get_int64 buf (arg * arg_size), get_int64 buf (arg * arg_size + 8) in 
+    vec128_of_int64s low high
+
+  let set_vec128 buf ~arg x = 
+    set_int64 buf (arg * arg_size) (vec128_low_int64 x); 
+    set_int64 buf ((arg * arg_size) + 8) (vec128_high_int64 x)
 
   let get_int32 t ~arg = get_int32 t (arg * arg_size)
   let get_int64 t ~arg = get_int64 t (arg * arg_size)
@@ -110,6 +125,7 @@ module Buffer = struct
     | Int64     -> get_int64
     | Nativeint -> get_nativeint
     | Float     -> get_float
+    | Vec128    -> get_vec128
 
   let set : type a. a typ -> t -> arg:int -> a -> unit = function
     | Int       -> set_int
@@ -117,6 +133,7 @@ module Buffer = struct
     | Int64     -> set_int64
     | Nativeint -> set_nativeint
     | Float     -> set_float
+    | Vec128    -> set_vec128
 
   (* This is almost a memcpy except that we use get/set which should
      ensure that the values in [dst] don't overflow. *)
@@ -161,6 +178,7 @@ let typ_size : type a. a typ -> int = function
   | Int64     -> 8
   | Nativeint -> Sys.word_size / 8
   | Float     -> 8
+  | Vec128    -> 16
 
 let rec sizes : type a. a proto -> int list = function
   | Ret typ         -> [typ_size typ]
